improve endpoint response
