+++ src/routers/events.ts
@@ -2,8 +2,10 @@ import type { PostgresJsDatabase } from 'drizzle-orm/postgres-js';
 import { default as express } from 'express';
 import type * as db from '../db';
 import { isLoggedIn } from '../middleware/isLoggedIn';
-import { getEvents, getRegistrationData, getRegistrationFields } from '../services/events';
+import { getEvents } from '../services/events';
 import { saveRegistration } from '../services/registrations';
+import { getRegistrationFields } from '../services/registrationFields';
+import { getRegistrationData } from '../services/registrationData';
 const router = express.Router();
 
 export function eventsRouter({ dbPool }: { dbPool: PostgresJsDatabase<typeof db> }) {
