+++ src/services/registrationFields.ts
@@ -1,0 +1,25 @@
+import type { PostgresJsDatabase } from 'drizzle-orm/postgres-js';
+import type { Request, Response } from 'express';
+import * as db from '../db';
+
+export function getRegistrationFields(dbPool: PostgresJsDatabase<typeof db>) {
+  return async function (req: Request, res: Response) {
+    const eventId = req.params.eventId;
+    if (!eventId) {
+      return res.status(400).json({ errors: ['eventId is required'] });
+    }
+
+    const event = await dbPool.query.events.findFirst({
+      with: {
+        registrationFields: {
+          with: {
+            registrationFieldOptions: true,
+          },
+        },
+      },
+      where: (fields, { eq }) => eq(fields.id, eventId),
+    });
+
+    return res.json({ data: event?.registrationFields });
+  };
+}
